requirements.txt
plaid-python==9.4.0
flask==1.1.2
